+++ spec/WFST.js
@@ -78,35 +78,35 @@ describe('WFST', function () {
     });
   });
 
-  describe('#save', function() {
-    it('should properly parse feature id', function() {
+  describe('#save', function () {
+    it('should properly parse feature id', function () {
       wfst.addLayer(layer);
       wfst.save();
 
       var transactionRequest = this.requests.pop();
       expect(transactionRequest.requestBody.indexOf('<wfs:Transaction')).to.be.equal(0);
       transactionRequest.respond(200, { "Content-Type": "text/xml" },
-      '<?xml version="1.0" encoding="UTF-8"?>' +
-      '<wfs:TransactionResponse xmlns:xs="http://www.w3.org/2001/XMLSchema"' +
-      ' xmlns:wfs="http://www.opengis.net/wfs"' +
-      ' xmlns:gml="http://www.opengis.net/gml"' +
-      ' xmlns:ogc="http://www.opengis.net/ogc"' +
-      ' xmlns:ows="http://www.opengis.net/ows"' +
-      ' xmlns:xlink="http://www.w3.org/1999/xlink"' +
-      ' xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="1.1.0"' +
-      ' xsi:schemaLocation="http://www.opengis.net/wfs http://localhost:8080/geoserver/schemas/wfs/1.1.0/wfs.xsd">' +
+        '<?xml version="1.0" encoding="UTF-8"?>' +
+        '<wfs:TransactionResponse xmlns:xs="http://www.w3.org/2001/XMLSchema"' +
+        ' xmlns:wfs="http://www.opengis.net/wfs"' +
+        ' xmlns:gml="http://www.opengis.net/gml"' +
+        ' xmlns:ogc="http://www.opengis.net/ogc"' +
+        ' xmlns:ows="http://www.opengis.net/ows"' +
+        ' xmlns:xlink="http://www.w3.org/1999/xlink"' +
+        ' xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="1.1.0"' +
+        ' xsi:schemaLocation="http://www.opengis.net/wfs http://localhost:8080/geoserver/schemas/wfs/1.1.0/wfs.xsd">' +
         '<wfs:TransactionSummary>' +
-          '<wfs:totalInserted>1</wfs:totalInserted>' +
-          '<wfs:totalUpdated>0</wfs:totalUpdated>' +
-          '<wfs:totalDeleted>0</wfs:totalDeleted>' +
+        '<wfs:totalInserted>1</wfs:totalInserted>' +
+        '<wfs:totalUpdated>0</wfs:totalUpdated>' +
+        '<wfs:totalDeleted>0</wfs:totalDeleted>' +
         '</wfs:TransactionSummary>' +
         '<wfs:TransactionResults/>' +
         '<wfs:InsertResults>' +
-          '<wfs:Feature>' +
-            '<ogc:FeatureId fid="test_feature_number.123"/>' +
-          '</wfs:Feature>' +
+        '<wfs:Feature>' +
+        '<ogc:FeatureId fid="test_feature_number.123"/>' +
+        '</wfs:Feature>' +
         '</wfs:InsertResults>' +
-      '</wfs:TransactionResponse>');
+        '</wfs:TransactionResponse>');
 
       var featureRequest = this.requests.pop();
       expect(featureRequest.requestBody.indexOf('<wfs:GetFeature')).to.be.equal(0);
@@ -114,8 +114,70 @@ describe('WFST', function () {
     });
   });
 
+  describe('#save xhr option withCredentials', function () {
+    var responseBody;
+
+    before(function () {
+      responseBody = '<?xml version="1.0" encoding="UTF-8"?>' +
+        '<wfs:TransactionResponse xmlns:xs="http://www.w3.org/2001/XMLSchema"' +
+        ' xmlns:wfs="http://www.opengis.net/wfs"' +
+        ' xmlns:gml="http://www.opengis.net/gml"' +
+        ' xmlns:ogc="http://www.opengis.net/ogc"' +
+        ' xmlns:ows="http://www.opengis.net/ows"' +
+        ' xmlns:xlink="http://www.w3.org/1999/xlink"' +
+        ' xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="1.1.0"' +
+        ' xsi:schemaLocation="http://www.opengis.net/wfs http://localhost:8080/geoserver/schemas/wfs/1.1.0/wfs.xsd">' +
+        '<wfs:TransactionSummary>' +
+        '<wfs:totalInserted>1</wfs:totalInserted>' +
+        '<wfs:totalUpdated>0</wfs:totalUpdated>' +
+        '<wfs:totalDeleted>0</wfs:totalDeleted>' +
+        '</wfs:TransactionSummary>' +
+        '<wfs:TransactionResults/>' +
+        '<wfs:InsertResults>' +
+        '<wfs:Feature>' +
+        '<ogc:FeatureId fid="test_feature_number.123"/>' +
+        '</wfs:Feature>' +
+        '</wfs:InsertResults>' +
+        '</wfs:TransactionResponse>';
+    });
+    it('save xhr option withCredentials true', function () {
+      wfst = new L.WFST({
+        typeNS: 'typeNS',
+        typeName: 'typeName',
+        namespaceUri: 'testuri',
+        withCredentials: true
+      });
+
+      wfst.addLayer(layer);
+      wfst.save();
+
+      var transactionRequest = this.requests.pop();
+      transactionRequest.respond(200, { "Content-Type": "text/xml" }, responseBody);
+
+      var featureRequest = this.requests.pop();
+      expect(featureRequest.withCredentials).to.be.equal(true);
+    });
+
+    it('save xhr option withCredentials false', function () {
+      wfst = new L.WFST({
+        typeNS: 'typeNS',
+        typeName: 'typeName',
+        namespaceUri: 'testuri',
+        withCredentials: false
+      });
+
+      wfst.addLayer(layer);
+      wfst.save();
+
+      var transactionRequest = this.requests.pop();
+      transactionRequest.respond(200, { "Content-Type": "text/xml" }, responseBody);
+
+      var featureRequest = this.requests.pop();
+      expect(featureRequest.withCredentials).to.be.equal(false);
+    });
+  });
+
   after(function () {
     xhr.restore();
   });
 });
-